+++ HISTORY.md
@@ -1,5 +1,6 @@
 0.5.0 [UNRELEASED]
 ===================
+- The main window is now resizable
 
 
 0.4.0